add left right camera based on toe-in? solution
