+++ config.py
@@ -1,4 +1,4 @@
-from camera_route import circle_local
+from camera_route import circle_local_left
 
 # Output figure store place
 pan1 = '.'
@@ -7,11 +7,14 @@ pan1 = '.'
 # map_fname = '/n/fink1/ggreen/bayestar/output/allsky_2MASS/compact/dust-map-3d-uncompressed.h5'#compact_10samp.h5'
 map_fname = 'data/AquilaSouthLarge2_unified.h5' 
 
+# Initiate processing core numbers
+n_procs = 10
+
 # Misc settings
 plot_props = {
     'fname': pan1 + '/3d/allsky_2MASS/AqS/AqS-loop-hq.png', #'3d/allsky_2MASS/grand-tour/simple-loop-att-v2-lq.png',
     'figsize': (10, 7),
-    'dpi': 10,
+    'dpi': 30,
     'n_averaged': 1,
     'gamma': 1.,
     'R': 3.1,
@@ -30,7 +33,8 @@ plot_props = {
 #camera_pos = paper_renderings()
 #camera_pos = Orion_flythrough(n_frames=400)
 #camera_pos = grand_tour_path(n_frames=20)#1600)
-camera_pos = circle_local(n_frames=1, l_0=30., b_0=5.)
+# camera_pos = circle_local(n_frames=3, l_0=30., b_0=5.)
+camera_pos = circle_local_left(n_frames=1, l_0=30., b_0=5.)
 #camera_pos = stereo_pair()
 
 #for key in camera_pos:
@@ -43,7 +47,7 @@ camera_props = {
     'n_x': 200*2,
     'n_y': 140*2,
     'n_z': 500*2,
-    'dr': 10./2,
+    'dr': 10./2,  # ???
     'z_0': 1., #(0., 0., 0.)
 }
