+++ render3d_test_mac_local.py
@@ -526,122 +526,12 @@ def gen_frame(mapper3d, camera_pos, camera_props,
     del img
 
 
-def Orion_flythrough(n_frames=10):
-    '''
-    Fly through the Orion ring.
-    '''
-    
-    l_0, b_0 = -148., -13.
-    a = (90. - b_0) * np.ones(n_frames, dtype='f8')
-    b = l_0 * np.ones(n_frames, dtype='f8')
-    
-    cl, sl = np.cos(np.radians(l_0)), np.sin(np.radians(l_0))
-    cb, sb = np.cos(np.radians(b_0)), np.sin(np.radians(b_0))
-    
-    d = np.linspace(0., 1000., n_frames)
-    x = d * cl * cb
-    y = d * sl * cb
-    z = d * sb
-    r = np.array([x, y, z]).T
-    
-    camera_pos = {
-        'xyz': r,
-        'alpha': a,
-        'beta': b
-    }
-    
-    return camera_pos
-
-
-def paper_renderings():
-    '''
-    A list of camera positions/orientations for
-    possible use in the paper.
-    '''
-    
-    r_0 = np.array([[  0.,  0.,  0.],
-                    [147., 26., 63.],
-                    [144., 41., 67.]])
-    a = np.array([ 96.0, 103.4, 104.1])
-    b = np.array([185.0, 190.2, 196.0])
-    
-    camera_pos = {
-        'xyz': r_0,
-        'alpha': a,
-        'beta': b
-    }
-    
-    return camera_pos
-
-
-def local_dust_path(n_frames=10):
-    '''
-    Construct camera path that:
-        * begins looking at the anticenter
-        * zooms back about 150 pc
-        * pans around by about 20 degrees
-    
-    This path focuses on the Orion molecular cloud complex,
-    Taurus, California and Perseus, i.e. most of the large
-    dust complexes in the Solar neighborhood.
-    '''
-    
-    A = 1.
-    
-    # Zoom out
-    r_0 = np.zeros((n_frames/4, 3), dtype='f8')
-    r_0[:,0] = np.linspace(0., A*150., r_0.shape[0])
-    r_0[:,2] = np.linspace(0., A*55., r_0.shape[0])
-    
-    dz = 20.
-    dR = 200.
-    
-    a_0 = 180./np.pi * np.arctan((r_0[-1,2] + dz) / (r_0[-1,0] + dR))
-    b_0 = 180.
-    
-    a_0 = np.linspace(90. + a_0/2., 90. + a_0, r_0.shape[0])
-    b_0 = b_0 * np.ones(r_0.shape[0])
-    
-    # Rotate around azimuthally, while bobbing in z
-    phi = 25. * np.pi/180. * np.sin(np.linspace(0., 2.*np.pi, int(3./4.*n_frames)))
-    theta = np.linspace(0., 2.*np.pi, phi.size)
-    #phi = np.linspace(0., 2.*np.pi, int(3./4.*n_frames))
-    R = r_0[-1,0]
-    Z = r_0[-1,2]
-    
-    r_1 = np.zeros((phi.size, 3), dtype='f8')
-    r_1[:,0] = R * np.cos(phi)
-    r_1[:,1] = R * np.sin(phi)
-    r_1[:,2] = Z + 20. * np.sin(theta)
-    #r_1[:,2] = Z * np.cos(phi/2.)
-    
-    a_1 = 90. + 180./np.pi * np.arctan((r_1[:,2] + dz) / (R + dR))
-    b_1 = np.mod(180. + 180./np.pi * phi, 360.)
-    
-    camera_pos = {'xyz': np.concatenate([r_0, r_1], axis=0),
-                  'alpha': np.hstack([a_0, a_1]),
-                  'beta': np.hstack([b_0, b_1])}
-    
-    #print camera_pos['alpha']
-    
-    return camera_pos
-
-def gen_movie():  
-
-    gen_movie_frames(map_fname, plot_props,
-                     camera_pos, camera_props,
-                     label_props, labels,
-                     n_procs=n_procs, verbose=True)
-
-
-
 def main():
     #grand_tour_path(n_frames=100)
     #circle_local()
-    from config import map_fname, plot_props, camera_props, label_props, camera_pos
+    from config import map_fname, plot_props, camera_props, label_props, camera_pos, n_procs
     
     # Generate frame
-    n_procs = 10
     n_procs = min([n_procs, len(camera_pos['alpha'])])
     
     # Points to project to camera coordinates
@@ -669,9 +559,7 @@ def main():
     gen_movie_frames(map_fname, plot_props,
                      camera_pos, camera_props,
                      label_props, labels,
-                     n_procs=n_procs, verbose=True)
-    # gen_movie()
-    
+                     n_procs=n_procs, verbose=True)    
     return 0
